Fix compiler errors on gcc 4.6.3 due to mismatched format strings and integer types.

+++ src/pagespeed/kernel/image/image_frame_interface.cc
@@ -94,7 +94,8 @@ GoogleString FrameSpec::ToString() const {
   return StringPrintf(
       "Frame: size %u x %u at (%u, %u) "
       "pixel_format: %d, duration_ms: %lu, disposal: %d, progressive: %s",
-      width, height, top, left, pixel_format, duration_ms, disposal,
+      width, height, top, left, pixel_format,
+      static_cast<unsigned long>(duration_ms), disposal,
       (hint_progressive ? "yes" : "no"));
 }
